polynomial: convert constructor to template and sort exponents
Change C arrays to ::std::array in order to make the templates
work.   Pull the templated methods into a separate header file because:

"The implementation of a non-specialized template must be visible to a
translation unit that uses it. The compiler must be able to see the
implementation in order to generate code for all specializations in
your code."