manipulate stack of generic pointers
Chapter 6, Exercise 20.

The exercise appears to be impossible to complete.  Perhaps that
was not so with earlier versions of C++ given that the Pohl book
dates to 1999.  Unlike with C, it is apparently not possible to
cast generic pointers back and forth to other types.  The errors

warning: pointer of type ‘void *’ used in arithmetic
error: ‘void*’ is not a pointer-to-object type

appear and compilation fails.

Task was : Employ a stack of generic pointers in order to reverse
an array of doubles.