SmarterStack: fix earlier inconsistencies and add subsequences
The earlier implementation of constructors, Pop() and Push() was
not internally consistent, so fix those problems.  Change the
tests to print to an ostringstream so that the << operator is
also tested.   Overload the function operator() to extract
subsequences from the stack, and exercise it by printing out
increasing subsequences.

Serves as solution to exercises 21 and 22 of Pohl chapter 6.