Create a class TermVector
The class TermVector is a vector of Terms that supports some of the simpler methods of
std::vector, but since it is backed by an array, is not currently
resizable.  It also lacks a Sort() method that Polynomial could use to
order Terms by exponent.

Part of the solution of Pohl Chapter 6, exercise 27.