Add Polynomial TermVector ctor and operator+
Binary operators must return by value a single object which is
constructible into the object returned by the operator.  Therefore
return a TermVector from the Polynomial operator+, as the termlist
constructor of Polynomial requires an object passed by reference, and
thus causes memory leaks.  A binary operator must not allocate memory
on the heap, as it returns an rvalue, not an lvalue.

Part of the solution of Pohl Chapter 6, exercise 27.