Add conditional debug mode to matrix functions
Add a DEBUG macro which, if defined, turns on assertions and printing
of values.  Parameterize the size of the matrices in the unit tests in
order to allow experimentation with the effect on run time. Fix a bug
in the transpose operation.

Result: turning on the DEBUG features more than doubles the run time
for 5x6 matrices.

Completes Exercise 29 of Pohl Chapter 6.