matrix: more unit tests and bug fixes
Make all the functions that process matrices work with index
offsets as demonstrated by corresponding passing tests.

This completes the solution for Exercise 31, Chapter 6 of Pohl,
and thus completes Chapter 6.  I'm sorely tempted to fully
rewrite class Matrix to replace Multiply() with operator*() and
copy semantics with move, and so forth, but instead I'll move on
to Chapter 7.