TemplatedStack: failed attempt to add constexpr ctor
The constexpr constructor that I copied from Philipp Schrader's span.h
worked fine in a unit test, but its presence in a class that includes
other constructors that allocate memory appears problematic.  No
matter what manipulations I tried, the end of a unit test resulted in
an attempt by the destructor needed by the other constructors to free
memory that the constexpr constructor had not allocated.