TemplatedStack: reverse() plus more ctors and tests
Add a true move constructor, a constructor that takes an array or an
array of R-values as an argument, and expressly delete the copy
constructor, which the compiler had a tendency to find during template
resolution.

Due to the Most Vexing
Parse (https://en.wikipedia.org/wiki/Most_vexing_parse), there appears
to be no way to use the default constructor, not even with
brace-initialization.  In order to create an empty stack, it may be
necessary to first create a full one and then empty it.

Given the inability to create an empty stack, the easiest way to
reverse() is perform what is essentially the push() within the array
constructor and then explicitly call pop().

The way to pass an array of R-value references is as an anonymous
array literal.  Note that the argument does not become an R-value
reference to the array, which is what I first anticipated.