macro-vs-template: compiler optimization and various inlines
The code represents the solution to Exercise 3 of Chapter 7 of Pohl,
which asks when executing code via a #define macro will produce a
different result than executing the same code via a template.  I added
a constexpr in-line implementation as well.  In fact, all three
implementations produce the same result either for a trivial or a
somewhat less trivial type.  The only difference among them is that,
with the almost-trivial Complex type, the array constructor is called
3 times, while with the template and constexpr implementations, the
array constructor is called only once:

    $ ./macro-vs-template
    macro: 140.296
    template: 140.296
    constexpr: 140.296

    macro: Array ctor
    Array ctor
    Array ctor
    152.897 + 2.01146i
    template: Array ctor
    152.897 + 2.01146i
    constexpr: Array ctor
    152.897 + 2.01146i

Apparently with the template and constexpr implementations, the
compiler reorders the cube() and sqrt() operations so that sqrt(),
which calls the array constructor, comes second.  With the macro
implementation, the compiler cannot perform the optimization.  There
could easily be other kinds of compiler optimizations that macros
defeat, showing the superiority of template and (when possible)
constexpr implementations.