template_cycle: rotate values in a list
Writing a cycle() function illustrates the notion that multiple
templates may be necessary in order to achieve a general-purpose
solution.  Simple assignment inside the function template definition
works fine for basic types and probably anything with a default
constructor and no specially defined assignment operator.  For a more
complex case like Polynomial, where the copy-assignment operator is
deleted, a different template that will move() rather than assign is
necessary.

For this program alone, the compiler could not find ::std::move
without the addition of "#include <utility>" in the header.  No other
program that uses ::std::move has this requirement, probably because
all others include iostream, which likely itself includes utility.

Also notable is that the test must specifically invoke the
template-instantiated functions (e.g. "tcycle::cycle") even though the
test is in the same namespace.

The code is the solution to Pohl Chapter 7, exercise 4.