template_vector: add equality test
Add a function which is effectively operator==() but is not an
overload since I couldn't figure out how to write a member function
whose type possibly differs from that of the object on which the
operator is called.   In other words, what I couldn't figure out was
how to declare

  bool operator==(const TemplateVector<U>& tv)

where U is possibly not a match to the class definition type T.  The
resulting function checks type equality via a call to ::std::is_same
that relies on constexpr to inform the compiler to check, not the
runtime.

I also realized that it's possible to overload a non-const getter with
a constant one.

This work is the beginning of exercise 8 in Chapter 7 of Pohl.