TemplateVector: assignment with type conversion
As part of Pohl Chapter 8, Exercise 8, create an assignment function
that uses static_cast to copy the values, but not the type, of the
contents of one TemplateVector to another.     The function must be a
friend of the TemplateVector class so that it can modify the contents,
but since the function needs two different template typenames in the
signature, it could not be declared inline, as I had done with the
operator<<() overload.

The so-defined function is an "extroverted" friend, as described at

https://stackoverflow.com/questions/4660123/overloading-friend-operator-for-template-class/4661372#4661372

so that any function which matches the signature could, unfortunately,
access the private variables, which is a security concern.

It turned out to be impossible to make a call to the function that
necessitated an impossible type conversion, as such a call fails at
compile-time, not at run-time.