TemplateVector: add in-place bisecting sort
Solution to Pohl Chapter 7, exercise 9.   I made little effort to make
the sort efficient since the point was to compare to qsort (AKA
::std::sort).  Typical results:

     Bisecting sort duration: 0.144 ms.
     ::std::sort duration: 0.051 ms.

     Bisecting sort duration: 0.079 ms.
     ::std::sort duration: 0.026 ms.

The bisecting sort typically has 3x the execution time of qsort.  An
earlier version of the bisecting sort that returned a reference to a
TemplateVector and then called the move() assignment operator took
another 3x more computation time.