template_list: use a map to store results from evaluating a list
Make a long std::list of random numbers, then use an ordered map to
count them.  I couldn't figure out how to declare the return value of
an insertion into the map using a template, so I tried "auto", which
worked.  I evaluated the code with the C++ preprocessor at
https://cppinsights.io/s/517ae3bb, which suggested that

::std::pair<std::map<T,int>>::iterator,bool>> result = countmap.insert(candidate);

should work, which it does not.  Line 13 of

http://www.cplusplus.com/reference/map/map/insert/

is also similar.

When the mean statistic was set to T, the calculation could overflow
in the intermediate results, as the values are summed before being
divided.  Perhaps there's a better way to prevent overflow without
undesirably converting type besides setting the mean to long double.

The included code is the solution to Pohl Exercise 12, Chapter 7.