templated list: inserter whose definition matches problem statement
Add a wrapper class for ::std::list in order to extend it with a
member function whose signature matches the problem statement of Pohl
Chapter 7, Exercise 14. A comment in template_list_impl.h explains
that the wrapper class is needed in order to add a function that
returns a templated type.  Add a new unit test for the new function.

Signed-off-by: Alison Chaiken <alison@she-devel.com>