template_largest: a function that takes an arbitrary sequence as a parameter
The first part of the solution to Pohl Chapter 7, exercise 16 which
states, "write an algorithm to find the second-largest element stored
in an arbitrary container.  Use STL containers vector<T>, list<T> and
set<T> . . ."  This version should work fine with any sequence
container, but likely will fail with associative containers like set.

My first attempts tried to return an iterator, but I could not figure
out how to write an iterator wrapper class that I could properly
initialize and pass in as a parameter from the test.