template_integrate: use generate() and accumulate()
The code represents the simplest solution to Pohl Chapter 7, Exercise
17.  Create a class that wraps the generator function, and make its call
operator() integrate the selected interval and evaluate the function to
be integrated.  Since the function object that is passed to generate()
cannot take any arguments, a new function object must be created for
each parameter-requiring function.  The function object constructor
could take as a parameter a function that itself take no parameters, but
it's hard to see why anyone would want to integrate such a function.