template_integrate: pass in pointer to function to be integrated
Instead of creating a new function object for each function to be
integrated, pass in a pointer to that function to do_fill(), which
replaces here the STL generate() function.  std::generate() will not
accept a function that takes an argument as a parameter.

The code is inspired by _A Tour of C++ 2nd Edition_ p. 84-86 and
https://stackoverflow.com/questions/2582161/function-pointer-as-parameter
and
https://stackoverflow.com/questions/877523/error-request-for-member-in-which-is-of-non-class-type.

This work completes Exercise 17 of Chapter 7 of Pohl, and with it,
Chapter 7.