template_integrate: provide lambda function as generator
A lambda function is a function object that is passed as an R-value on
the stack.  As such, it is not possible to store it in the
function-pointer member of the FunctionWrapper class.  Instead of
rewriting ::std::generate() as a class method, therefore, simply write
a non-member method that takes a ::std::function as a parameter.

Providing a typename as a parameter that helps define the iterator and
function parameters does not work.  Thus the use of ForwardIterator,
as ::std::vector<T>::iterator does not work, nor does
::std::function<T(T)>.