reverse_list: reverse a std::list by means of its iterators
I've heard that reversing a std::list is a common interview question,
so I decided to take a crack at it while finishing up the Pohl
templates chapter.

The solution includes a public-API function that takes only a list as
a parameter, and an internal function that takes parameters that
facilitate recursion, which is natural for lists.  Four parameters to
this function seems too many, so I tried to store them in a struct,
but could not get the template substitution to work.  Defining

template <typename ForwardIterator, typename ReverseIterator> struct
ListProperties

simply failed template deduction, and then an attempt with decltype(decval)

+  ListProperties<
+      decltype(::std::declval<typename ::std::list<T>::iterator>()),
+      decltype(::std::declval<typename ::std::list<T>::reverse_iterator>())>
+      lp(alist.begin(), alist.rbegin());
+  lp.total_swaps = alist.size() / 2;
+  lp.swaps = 0U;
+  ReverseInternal(lp);

failed because declval generates an R-value reference that then could
not be stored in the struct for use by ReverseInternal.  Apparently
the decltype(declval) construct is useful for passing in methods that
will be immediately executed, not for passing in class members that a
constructor will store.