student_inheritance: inheritance complicates friend functions
Friend functions cannot be virtual or be explicitly invoked via
membership qualification since they are not class members.  Since unit
tests that invoke the extraction operator need strings to test, it's
usual to override operator<<() with a friend and make it return an
ostringstream.  Having identical function signatures for the friends
of a base and parent class therefore clearly presents problems which I
sought to solve by making one extraction operator return an ostream
and the other an ostringstream explicitly.  That proved to present its
own confusing difficulties, as ostringstream inherits from ostream!

This effort represents only the first part of Exercise 1, Chapter 8 of
Pohl.