student_inheritance: effect of casting on override resolution
I created print() methods for the base and derived class that have
matching function signatures by having them invoke the extraction
operators.  These matching-signature functions are then used to study
the override resolution of various casts.  Notably, direct invocation
of the derived class' method from a pointer to the base class fails
compilation, and casting a base object to a derived pointer results in
a allocation failure when the derived function attempts to obtain a
memory chunk corresponding to the random 64-bit sizes of the
uninitialized members of the derived object.

This work satisfies Exercise 2 of Pohl, Chapter 8.