student_inheritance: derived cannot modify parent members
By default, changes to the members of a parent class are not modified
by changes inside a derived object unless the parent's member is
explicitly also a member of the derived class.

Add some typeid() checking and some assertions about string size() in
order to prevent failures from unwise casting, but they do not
work. The results are documented in inheritance_casting_main.txt.

This work fulfills Exercise 3, Chapter 8 of Pohl.  Notably the text
says, "The declaration of an identifier in an outer scope hides all
declarations of that identifier in outer scopes.  A base class is an
outer scope of any class derived from it.  This rule is independent of
whether the names are declared virtual.  Access restrictions (private,
protected) are orthogonal to function selection.  If the selected
function is inaccessible, that is a compile-time error."