async_logger: fix broken tests
I realized that TearDown() is never called if the function signature
is altered to take a parameter.  In other words, TearDown() cannot be
overloaded.

When I changed the code so that TearDown() was called, the test locked
up.   That turned out to be because the test destructor is called
after TearDown(), which means that the new lock I added was never
released.  Since SetUp() is called after the test's constructor, this
all makes sense.

The new design presented here moves the code from TearDown() to the
test destructor and, happily, dispenses with the added lock.