multiple_inheritance: improve extraction operators and fix tests
The existing unit tests failed to properly exercise the extraction
operators, which in fact did not work as intended.   Earlier code made
extraction operators friends of the classes with which they were
associated in order that tests could provide an ostringstream rather
than an ostream as one of the parameters.   The async_logger code
showed a better way, where the usual operator<<() syntax is preserved,
but the test receives the output as a parsable string.   Change to
that approach here and add more tests.

There is no extraction operator yet for StudentWorker, which requires
preventing multiple execution of the base Person class extraction
operator.