multiple_inheritance: prevent repeated base class operator execution
There appears to be no method to detect if characters have been
written to an ostream, in contrast to an istream, which has a notion
of emptiness.   The only technique I could therefore find to indicate
whether the base Person class extraction operator had already run was
to create a boolean flag.   The flag prevents Person's extraction
operator from running twice when StudentWorker calls both Student's
and Worker's extraction operators.