multiple_inheritance: read Persons from file, push onto lists
Read Person data from a file one line at a time.  Provide safe
string-conversion functions that fail gracefully for numeric data.
Create lists of shared_ptrs to Persons so that there is no naked "new"
in the list creation code, and so that the runtime deletes the Persons
when the code completes.  Access the lists of the four Person types
via an array for convenience. Move the operator<<() definitions from
the header to the library.  Provide tests for Person input data
processing and list population.