template_vector: main function allows user to correct errors
Let the user provide a new parameter value when an invalid one causes
a function to throw an exception.  At the same time, exercise wide
characters just to see how they work.

These last two commits represent the work for Exercise 1-3 of Pohl
Chapter 9.  The actual Exercises 1 and 2 present a bubble sort with an
off-by-one error that I can't bear to run.