polynomial: bring to 100% coverage
Add some new operators and functions to class Polynomial, which was
both incomplete and a bit broken.  In particular, the methods
permitted the existence of empty Terms, defined as those whose
coefficient was zero.  Add a function to delete them and the whole
Polynomial should it prove to be entirely empty after an addition or
subtraction operation.  Add previously missing tests as well as new
one to exercise the new operators.