multiple_inheritance: improve file I/0 tests
Bring multiple_inheritance_lib.cc and multiple_inheritance.h to 100%
coverage.  gcov fails to properly account for DEATH tests and still
under-reports the coverage.  Totally rewrite the file I/O error
handler to account for getline() failures as well as opening errors.
Remove a couple of lines of code that cannot be reached due to earlier
failures in execution paths.

lcov still shows zero coverage on student_inheritance.h and
student_inheritance_lib.cc.  It's not clear that there a solution to
this problem.