take matrix_lib to nearly 100% coverage
Cleaned up matrix_lib extensively.  Made accessors for array bounds
easier to understand.  Added multiple tests.  Everything but an
illegal enum value is now covered and there appears to be no way to
generate such a value without triggering one of the sanitizers.
Remove unreachable code and decide not to support characteristic
polynomial computation for larger matrices.  It's a lot of work to do
so, and this project is about C++, not matrix manipulation.

As usual, LCOV doesn't understand DEATH tests, so the coverage numbers
are some erroneously low.