Attempt to create custom operator new to test alloc failure
Two failed attempts to override operator new in order to test the paths
in SmarterStack following the code in

https://en.cppreference.com/w/cpp/memory/new/operator_new#Class-specific_overloads

Notably those examples only invoke the replacement operator from a
main() function, not from a method of the class where the new operator
is defined.   The compiler reports no matching function signature if the
constructors that allocate attempt to call the version of new that takes
a bool as a second argument, whereas such a call works fine from a test,
but then throws without calling SmarterStack's constructor.   The
original intention of adding test coverage then fails, making this
exercise pointless.

Note that the example code in cppreference only compiles with C++17.