Makefile: create a generic compilation rule for non-templated tests
All non-templated tests can be compiled with one generic rule.
Creation of the rule greatly reduces duplication in the Makefile.  A
key point is the that "$^" automatic variable is needed for the
compilation, which matches all prerequisites as explained in

https://www.gnu.org/software/make/manual/make.html#Automatic-Variables

Most rules depend only on one CC file, in which case they can instead
compile "$<".

Prerequisites are listed after the colon following the target name.
Listing unnecessary headers as prequisites has the effect of causing
the target to be recompiled if the headers change, but it does not
affect the executed compilation command.