dyn_string: convert to smart pointers
Get rid of operator new in favor of std::unique_ptr.  Replace the copy
constructor with a move variant, and create a new test for it.  Change
all occurrences of strcpy() to strncpy() and all of strcat() to
strncat().   Get rid of an unnecessary copy in the string-reversal
function.

The code is at 100% coverage, but that appears to because gcov doesn't
require testing of failed assertions.  Replacing the assertions with
error handlers would reduce the coverage count without the addition of
more tests.