template_list: pass data to helper class with shared_ptr
Change raw pointers to references where possible.  Passing a reference
to the constructor of the ListHelper class results in a copy of the
contents of passed std::list, meaning that any changes made to the
ListHelper's list member do not propagate to the caller's list.
Resolve this problem by passing a shared_ptr rather than a reference.
Improve test coverage to 100%.