gcc -fanalyzer: not yet usable for C++
GCC 10.2.1 as shipped by Debian does not include support for "-fanalyzer", as
the "Configured with" output of

    $ /usr/bin/g++ -v

shows.   GCC git is at version 12.0.1.   Compiling it with

    $ apt-get install libmpfr6 libmpfr-dev libmpfrc++-dev
    $ apt-get install libmpc3 libmpc-dev
    $ make distclean
    $ ./configure --disable-multilib --enable-languages=c,c++ --enable-analyzer
    $ make

produces a compiler with analyzer support, but it does not yet work well for
C++, as reported at

    https://gcc.gnu.org/bugzilla/showdependencytree.cgi?id=97110

In particular, the analyzer does not ignore third-party headers from googletest,
which dominate the output.