polynomial_lib: ignore clang-tidy warning about nullptr dereference
clang-tidy flags

     while (*ap == *bp)

in polynomial_lib's operator==() as a potential source of a nullptr
dereference.   There is no danger of a nullptr dereference as polyonomial_lib's
operator==() calls term_lib's operator==(), which returns false if the next
pointers of two Terms differ.   The result is that the previous conditional

   (*(a.h_.get()) == *(b.h_.get()))

is false.   Code that placates clang-tidy

    if ((!ap) || (!bp)) {
        return false;
    }

is therefore unreachable.