Fix more errors identified by clang-tidy and valgrind
Get rid of strcpy() in favor of strncpy().  Fix uninitialized values and removed
undefined behavior.

Use NOLINTNEXTLINE to ignore clang-tidy's claim that reading the first member of
a list in multiple_inheritance_lib.cc has "side effects.  Also ignore the claim
that a comparison with end() value of an iterator is illegal because it "is a
garbage value."   Yes, we know!

Delete one more useless library.   Standardize the name of a header.