find min sum of values chosen from unique rows and columns of matrix
Minimize the sum of values chosen from an NxN cost matrix given the constraint
that exactly one value must be selected from each row and column of the matrix.
Represent potential solutions as an N-vector where the ith value designates the
column from which a cost in the the ith row is chosen.

The problem was stated in an issue of an ACCU journal that I have forgotten.
The one-from-each-row and one-from-each-column constraints are the same as for
the famous "nine queens" problem.  There undoubtedly is a better recursive
solution which I will endeavor to discover next.

In order to find 12 unique sums for the 4x4 input, the code considered 48 mask
possibilities, which is 2x the total number of unique permutations.

In order to find 4 unique sums for the 3x3 matrix, the code considered 18
possibilities, which is 3x the total number of unique permutations.

The fraction of duplicate cost sums increases with the count of pairs of entries
in the cost matrix which add up to the same value.  All the possible masks would
have unique sums if the numbers in the cost matrix were sufficiently different
from one another.