Makefile: allow users to prepend additional C++ flags from cmdline
The keyword "override" in a Makefile variable assignment means that
the Makefile-specified values will be applied no matter what is typed
on the cmdline.  "override" in combination with "+=" means that the
user is free to add additional flags but not to completely redefine
the variable. "override" without "+=" means that cmdline-supplied
flags are ignored.  The combination makes obvious sense when code will
not will not compile without the Makefile-set values.