multiple_inheritance: remove broken source code
There appears to be no way to make the code compile with GCC 13.3
without completing rewriting it.   Compilation fails with an error stating that
class Person is an incomplete type.   Apparently methods whose argument is a
subclass of Person cannot manipulate containers which hold the Person type.

I tried the advice at

https://stackoverflow.com/questions/13594549/base-class-has-incomplete-type

and attempted to write Person class member overloads for operator[], but met
with no success.   Since as far as I can tell, multiple inheritance is of no
interest in C++, I'm just deleting the code.