condvar-stl: speed-up with notify_all() and try_lock()
Make the comparison between condvar-pthread and condvar-stl versions
fairer by changing the STL version to try_lock() instead of a blocking
lock.  Also change the timeout version to sending a broadcast wakeup
rather than signalling a single, possibly wrong waiter.  Remarkably,
the change to the signalling makes about a factor of 3 difference in
speed.  Why condvar-stl is so much faster with a broadcast signal in
the timeout version when the condvar-pthread version is not is
unclear.